Use and-let* even more
The previous commit also replaces many instances of `when-let' with
`and-let'.  In those cases I did not hesitate.  This commit addresses
instances where side-effects are involved.

My current reasoning is that if the returned value matters, then
`and'/`and-let*' should be used, and whether there are side-effects is
only a secondary consideration.  Of course there is a gray area; if it
is really important to emphasize the side-effect, then one might still
want to use `when'/`when-let*'.

The simplest case is the change to `magit-diff-visit--hunk'.  That
function doesn't actually visit anything; instead it returns the hunk
which is relevant when visiting.  But the function name could give the
false impression that it somehow visits a hunk.  Preventing that false
impression is a good reason to use `and-let*'.  Additionally the
side-effect of the first form in BODY is local to the BODY, to the
outside code this is an irrelevant implementation detail.

`magit-diff--dwim' call `deactivate-mark', which clearly is a
side-effect that affects the global buffer state.  However this
side-effect is not relevant to the direct outside code, which on the
other hand cares very much about the returned value.

Finally `magit-hunk-goto-successor', which is very much about the
side-effect of "going to the successor".  This is so obvious that I
don't think it is necessary to signal the side-effect-ness.  Again the
returned value is used for flow-control, which should be made obvious,
so `and-let*' it is.